chore(1-3336): more renaming
